requirements.txt
fastapi
uvicorn
piper-tts
numpy
sounddevice
